app/api/requirements.txt
Flask==2.3.0
psycopg2-binary==2.9.6
Flask-CORS==4.0.0